:recycle: useMathField로 리팩토링
